CHORE refactored checkConnectivity function
