+++ Makefile
@@ -5,15 +5,15 @@ RELEASE_DIR = release
 
 # Default build for current architecture
 build: build-arm build-amd32
-	go build -o my-incident-checker main.go
+	go build -o my-incident-checker
 
 # Build for ARM (e.g., Raspberry Pi)
 build-arm:
-	GOOS=linux GOARCH=arm go build -o my-incident-checker-arm main.go
+	GOOS=linux GOARCH=arm go build -o my-incident-checker-arm
 
 # Build for 32-bit AMD/Intel
 build-amd32:
-	GOOS=linux GOARCH=386 go build -o my-incident-checker-386 main.go
+	GOOS=linux GOARCH=386 go build -o my-incident-checker-386
 
 test:
 	go test ./...
