+++ main.go
@@ -58,12 +58,10 @@ func NewLogger() (*Logger, error) {
 const (
 	notificationEndpoint = "https://ntfy.sh/dapidi_alerts"
 	incidentsEndpoint    = "https://status-api.joseserver.com/incidents/recent?count=10"
-	connectivityCheck    = "https://www.google.com"
 	heartbeatEndpoint    = "https://nosnch.in/2b7bdbea9e"
 	pollInterval         = 5 * time.Second
 	heartbeatInterval    = 5 * time.Minute
 	timeFormat           = "2006-01-02T15:04:05.999999"
-	connectTimeout       = 10 * time.Second
 
 	serialPort = "/dev/ttyUSB0" // Change to the serial/COM port of the tower light
 	baudRate   = 9600
@@ -198,24 +196,6 @@ func (l *Light) Clear() error {
 			return err
 		}
 	}
-	return nil
-}
-
-func checkConnectivity() error {
-	client := &http.Client{
-		Timeout: connectTimeout,
-	}
-
-	resp, err := client.Get(connectivityCheck)
-	if err != nil {
-		return fmt.Errorf("connectivity check failed: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("connectivity check failed with status code: %d", resp.StatusCode)
-	}
-
 	return nil
 }
